# docs/swagger.yaml
basePath: /v1
definitions:
...
  main.CreateUserTokenPayload:
    properties:
      email:
        maxLength: 255
        type: string
      password:
        maxLength: 72
        minLength: 3
        type: string
    required:
    - email
    - password
    type: object
  main.GoogleCallbackPayload:
    properties:
      code:
        type: string
      state:
        type: string
    required:
    - code
    - state
    type: object
  main.GoogleLoginResponse:
    properties:
      auth_url:
        type: string
      state:
        type: string
    type: object
  main.RegisterUserPayload:
    properties:
      email:
        maxLength: 255
        type: string
      first_name:
        maxLength: 255
        type: string
      last_name:
        maxLength: 255
        type: string
      password:
        maxLength: 72
        minLength: 3
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  main.ResendConfirmationPayload:
    properties:
      email:
        maxLength: 255
        type: string
    required:
    - email
    type: object
...
  main.UserWithToken:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      google_id:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_name:
        type: string
      token:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for Restaurant Employee Scheduling Application (RESA)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: RESA API
paths:
  /authentication/google:
    post:
      consumes:
      - application/json
      description: Generates and returns the Google OAuth authorization URL
      produces:
      - application/json
      responses:
        "200":
          description: OAuth URL and state token
          schema:
            $ref: '#/definitions/main.GoogleLoginResponse'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Initiates Google OAuth login
      tags:
      - authentication
  /authentication/google/callback:
    post:
      consumes:
      - application/json
      description: Exchanges authorization code for user info and creates/authenticates
        user
      parameters:
      - description: OAuth callback data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.GoogleCallbackPayload'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Handles Google OAuth callback
      tags:
      - authentication
  /authentication/refresh:
    post:
      consumes:
      - application/json
      description: Creates a new token with a fresh expiry time using an existing
        valid token
      produces:
      - application/json
      responses:
        "200":
          description: New token
          schema:
            type: string
        "401":
          description: Unauthorized - invalid or expired token
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Refreshes an authentication token
      tags:
      - authentication
  /authentication/resend-confirmation:
    post:
      consumes:
      - application/json
      description: Resends the confirmation email to an inactive user
      parameters:
      - description: User email
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.ResendConfirmationPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Resends confirmation email
      tags:
      - authentication
  /authentication/token:
    post:
      consumes:
      - application/json
      description: creates a token for a user
      parameters:
      - description: User credentials
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.CreateUserTokenPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Token
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Creates a token
      tags:
      - authentication
  /authentication/user:
    post:
      consumes:
      - application/json
      description: Registers a user
      parameters:
      - description: User credentials
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.RegisterUserPayload'
      produces:
      - application/json
      responses:
        "201":
          description: User registered
          schema:
            $ref: '#/definitions/main.UserWithToken'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Registers a user
      tags:
      - authentication
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"